+++ bc/cases/tests/test_models.py
@@ -1,6 +1,7 @@
 import textwrap
 from unittest import mock
 
+from django.contrib.sessions.models import Session
 from django.http.response import HttpResponse
 from django.test import TestCase
 
@@ -20,7 +21,9 @@ class CaseFormPageTest(TestCase):
     @mock.patch("bc.cases.models.get_client")
     def test_page_loads_when_client_not_configured(self, mock_get_client):
         mock_get_client.side_effect = RespondClientException
-        resp = self.client.get(self.case_form_page.url)
+        resp = self.client.get(
+            self.case_form_page.url + self.case_form_page.reverse_subpage("form_route")
+        )
         self.assertEqual(resp.status_code, 200)
 
     @mock.patch("bc.cases.models.get_client")
@@ -46,8 +49,15 @@ class CaseFormPageTest(TestCase):
         mock_get_client.return_value = mock.Mock(
             **{"create_case.return_value": HttpResponse(response_xml)}
         )
+        resp = self.client.post(
+            self.case_form_page.url + self.case_form_page.reverse_subpage("form_route")
+        )
+        self.assertRedirects(
+            resp, self.case_form_page.url, fetch_redirect_response=False
+        )
+
         with self.assertTemplateUsed("patterns/pages/cases/form_page_landing.html"):
-            resp = self.client.post(self.case_form_page.url)
+            resp = self.client.get(self.case_form_page.url)
         self.assertEqual(resp.status_code, 200)
         self.assertContains(resp, "10028")
         self.assertNotContains(resp, "Response Xml Test")
@@ -74,13 +84,67 @@ class CaseFormPageTest(TestCase):
         mock_get_client.return_value = mock.Mock(
             **{"create_case.return_value": HttpResponse(response_xml)}
         )
+
+        resp = self.client.post(
+            self.case_form_page.url + self.case_form_page.reverse_subpage("form_route")
+        )
+        self.assertRedirects(
+            resp, self.case_form_page.url, fetch_redirect_response=False
+        )
+
+        expected_case_reference = "DIS 10028"
+        session = Session.objects.get(session_key=self.client.session.session_key)
+        self.assertEqual(
+            session.get_decoded()[self.case_form_page.get_case_reference_session_key()],
+            expected_case_reference,
+        )
+
         with self.assertTemplateUsed("patterns/pages/cases/form_page_landing.html"):
-            resp = self.client.post(self.case_form_page.url)
+            resp = self.client.get(self.case_form_page.url)
         self.assertEqual(resp.status_code, 200)
         self.assertContains(resp, "10028")
         self.assertNotContains(resp, "Response Xml Test")
-        expected_case_reference = "DIS 10028"
         self.assertEqual(resp.context["case_reference"], expected_case_reference)
+
+    @mock.patch("bc.cases.models.get_client")
+    @mock.patch("bc.cases.models.ApteanRespondCaseFormPage.get_form")
+    def test_good_submission(self, mock_get_form, mock_get_client):
+        response_xml = textwrap.dedent(
+            """\
+            <?xml version="1.0" encoding="utf-8"?> <caseResponse version="2" xmlns:="http://www.aptean.com/respond/caseresponse/2">
+            <case Id="01fc3664f6194732a37f61222cca21bd" Name="DIS - 10028 Response Xml Test," Tag="">
+            </case>
+            </caseResponse>
+        """  # noqa
+        )
+        mock_get_client.return_value = mock.Mock(
+            **{"create_case.return_value": HttpResponse(response_xml)}
+        )
+        resp = self.client.post(
+            self.case_form_page.url + self.case_form_page.reverse_subpage("form_route")
+        )
+        with self.subTest("Good submission redirects to index route"):
+            self.assertRedirects(
+                resp, self.case_form_page.url, fetch_redirect_response=False
+            )
+
+        with self.subTest("Session has 'show landing page' option set"):
+            session = Session.objects.get(session_key=self.client.session.session_key)
+            self.assertTrue(
+                session.get_decoded()[
+                    self.case_form_page.get_landing_page_session_key()
+                ]
+            )
+
+    @mock.patch("bc.cases.models.get_client")
+    @mock.patch("bc.cases.models.ApteanRespondCaseFormPage.get_form")
+    def test_success_page(self, mock_get_form, mock_get_client):
+        session = self.client.session
+        session[self.case_form_page.get_landing_page_session_key()] = True
+        session.save()
+        with self.assertTemplateUsed("patterns/pages/cases/form_page_landing.html"):
+            resp = self.client.get(self.case_form_page.url)
+        self.assertNotIn("form", resp.context)
 
 
 class CaseNameFormattingTest(TestCase):